-create splash screen android & ios. -screen util done.
