+++ android/app/src/main/res/drawable-mdpi/splash.png
