+++ flutter_native_splash.yaml
@@ -1,0 +1,9 @@
+flutter_native_splash:
+  color: "#ffffff"
+  image: assets/images/splash_ios_android_11.png
+
+
+  android_12:
+    image: assets/images/splash_android_12.png
+    color: "#ffffff"
+
