+++ lib/main_admin.dart
@@ -2,7 +2,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:vista_market/firebase_options.dart';
 import 'package:vista_market/src/utils/env.dart';
-import 'package:vista_market/src/utils/vista_market_app.dart';
+import 'package:vista_market/src/vista_market_app.dart';
