+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   flutter_screenutil: ^5.9.3
   flutter_secure_storage: ^9.2.2
   connectivity_plus: ^6.0.5
+  flutter_native_splash: ^2.4.1
 
 dev_dependencies:
   flutter_test:
