+++ test/widget_test.dart
@@ -7,7 +7,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:vista_market/src/utils/vista_market_app.dart';
+import 'package:vista_market/src/vista_market_app.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
